feat: Payment를 FinishedEvent로 rename
